Refactor: /users mutations
