+++ backend/src/core/users/users.service.ts
@@ -332,6 +332,10 @@ export class UsersService implements OnModuleInit {
     return {
       ...created,
       isPersonal: false as const,
+      user: {
+        uid: callingUser.uid,
+        email: callingUser.email,
+      },
     };
   }
