+++ frontend/pages/organizations/accept-invite.tsx
@@ -7,14 +7,15 @@ import { Message } from '@/components/lib/Message';
 import { Spinner } from '@/components/lib/Spinner';
 import { useAuth } from '@/hooks/auth';
 import { useOrganizationsLayout } from '@/hooks/layouts';
+import { useMutation } from '@/hooks/mutation';
 import type { ParsedError } from '@/hooks/organizations';
-import { useAcceptOrgInvite } from '@/hooks/organizations';
+import { mutateOrganizations } from '@/hooks/organizations';
 import { StableId } from '@/utils/stable-ids';
 import type { NextPageWithLayout } from '@/utils/types';
 
 const AcceptOrgInvite: NextPageWithLayout = () => {
   const router = useRouter();
-  const acceptMutation = useAcceptOrgInvite();
+  const acceptMutation = useMutation('/users/acceptOrgInvite', { onSuccess: () => mutateOrganizations() });
   const { identity } = useAuth();
   const queryToken = router.query.token;
   const token = Array.isArray(queryToken) ? queryToken[0] : queryToken;
